+++ src/Corelai.Prime.App/src/app/main.tsx
@@ -2,10 +2,10 @@ import {StrictMode} from 'react'
 import {createRoot} from 'react-dom/client'
 import '../styles/main.css'
 import App from './App.tsx'
-import {loadSettings} from "../utils/appSettings.ts";
-import {SettingsProvider} from "../context/SettingsContext.tsx";
+import {AppSettings, loadSettings} from "../utils/appSettings.ts";
+import SettingsProvider from "../providers/SettingsProvider.tsx";
 
-loadSettings().then((settings) => {
+loadSettings().then((settings:AppSettings) => {
 
     createRoot(document.getElementById('root')!).render(
         <StrictMode>
